+++ src/base/widget.h
@@ -26,6 +26,7 @@
 #include "tkc/mem.h"
 #include "tkc/wstr.h"
 #include "tkc/value.h"
+#include "tkc/darray.h"
 #include "tkc/rect.h"
 #include "tkc/emitter.h"
 
@@ -65,9 +66,8 @@ typedef ret_t (*widget_set_prop_t)(widget_t* widget, const char* name, const val
 typedef widget_t* (*widget_find_target_t)(widget_t* widget, xy_t x, xy_t y);
 typedef widget_t* (*widget_create_t)(widget_t* parent, xy_t x, xy_t y, wh_t w, wh_t h);
 typedef ret_t (*widget_on_destroy_t)(widget_t* widget);
-typedef ret_t (*widget_recycle_t)(widget_t* widget);
 
-typedef struct _widget_vtable_t {
+struct _widget_vtable_t {
   uint32_t size;
   const char* type;
   /*克隆widget时需要复制的属性*/
@@ -91,6 +91,11 @@ typedef struct _widget_vtable_t {
    */
   uint32_t is_keyboard : 1;
 
+  /**
+   * 是否启用pool(如果启用pool，控件需要对其成员全部变量初始化，不能假定成员变量为0)。
+   */
+  uint32_t enable_pool : 1;
+
   widget_create_t create;
   widget_get_prop_t get_prop;
   widget_get_prop_default_value_t get_prop_default_value;
@@ -112,9 +117,8 @@ typedef struct _widget_vtable_t {
   widget_on_remove_child_t on_remove_child;
   widget_on_event_t on_event;
   widget_find_target_t find_target;
-  widget_recycle_t recycle;
   widget_on_destroy_t on_destroy;
-} widget_vtable_t;
+};
 
 /**
  * @class widget_t
@@ -331,11 +335,11 @@ struct _widget_t {
    */
   widget_t* key_target;
   /**
-   * @property {array_t*} children
+   * @property {darray_t*} children
    * @annotation ["readable"]
    * 全部子控件。
    */
-  array_t* children;
+  darray_t* children;
   /**
    * @property {emitter_t*} emitter
    * @annotation ["readable"]
@@ -1521,6 +1525,10 @@ ret_t widget_re_translate_text(widget_t* widget);
 /**
  * @method widget_init
  * 初始化控件。仅在子类控件构造函数中使用。
+ *
+ * > 请用widget\_create代替本函数。
+ *
+ * @depreated
  * @annotation ["private"]
  * @param {widget_t*} widget widget对象。
  * @param {widget_t*} parent widget的父控件。
@@ -1534,6 +1542,22 @@ ret_t widget_re_translate_text(widget_t* widget);
  */
 widget_t* widget_init(widget_t* widget, widget_t* parent, const widget_vtable_t* vt, xy_t x, xy_t y,
                       wh_t w, wh_t h);
+
+/**
+ * @method widget_create
+ * 创建控件。仅在子类控件构造函数中使用。
+ * @annotation ["private"]
+ * @param {widget_t*} parent widget的父控件。
+ * @param {widget_vtable_t*} vt 虚表。
+ * @param {xy_t}   x x坐标
+ * @param {xy_t}   y y坐标
+ * @param {wh_t}   w 宽度
+ * @param {wh_t}   h 高度
+ *
+ * @return {widget_t*} widget对象本身。
+ */
+widget_t* widget_create(widget_t* parent, const widget_vtable_t* vt, xy_t x, xy_t y, wh_t w,
+                        wh_t h);
 
 /**
  * @method widget_update_style